Create Tests/functionsValidatePasswordTest.php
This is to check and prove the functionality of the code https://en.wikipedia.org/wiki/Unit_testing 
It helps keep code working when you make changes. 
Next time when a change is made to validate_password() this file can be run with phpunit to make sure the function still works with the existing code.