Treat SA tag and refuse scores properly
sa_tag and sa_refuse settings should be treated as minimum values to induce the effect, not as maximum values to NOT induce it.
For example, if I have `sa_tag=20`, I expect all messages with spam score 2.0 or above tagged, as opposed to only messages above 2.0 being tagged.
Fixes #223